chore(api): move endpoint `/api/poll` to `/api/poll/:pollId`
  ## what
  - move endpoint `/api/poll` to `/api/poll/:pollId`

  ## how

  ## why
  - this required because of the way polling works in the upstream api
  - the first time when polling the polling ID would be 0
  - after that the polling ID would be the latest `submission ID`

  ## where
  - ./src/app/api/poll/route.ts -> ./src/app/api/poll/[pollId]/route.ts

  ## usage
